+++ pkg/chaosdaemon/setup_test.go
@@ -18,15 +18,12 @@ package chaosdaemon
 import (
 	"testing"
 
-	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"sigs.k8s.io/controller-runtime/pkg/envtest/printer"
 )
 
 func TestAPIs(t *testing.T) {
 	RegisterFailHandler(Fail)
 
-	RunSpecsWithDefaultAndCustomReporters(t,
-		"chaosdaemon Suite",
-		[]Reporter{printer.NewlineReporter{}})
+	RunSpecs(t, "chaosdaemon Suite")
 }